+++ src/conversion.ts
@@ -1,4 +1,4 @@
-// Katakana -> Hiragana conversion tables
+// Katakana / decomposed hiragana -> Composed hiragana conversion tables
 
 // prettier-ignore
 const HANKAKU_KATAKANA_TO_HIRAGANA = [
@@ -10,19 +10,28 @@ const HANKAKU_KATAKANA_TO_HIRAGANA = [
   0x3081, 0x3082, 0x3084, 0x3086, 0x3088, 0x3089, 0x308a, 0x308b, 0x308c,
   0x308d, 0x308f, 0x3093,
 ];
-// prettier-ignore
-const VOICED_KATAKANA_TO_HIRAGANA = [
-  0x30f4, 0xff74, 0xff75, 0x304c, 0x304e, 0x3050, 0x3052, 0x3054, 0x3056,
-  0x3058, 0x305a, 0x305c, 0x305e, 0x3060, 0x3062, 0x3065, 0x3067, 0x3069,
-  0xff85, 0xff86, 0xff87, 0xff88, 0xff89, 0x3070, 0x3073, 0x3076, 0x3079,
-  0x307c,
-];
-// prettier-ignore
-const SEMIVOICED_KATAKANA_TO_HIRAGANA = [
-  0x3071, 0x3074, 0x3077, 0x307a, 0x307d
-];
 
-// Half & full-width katakana to hiragana conversion
+// prettier-ignore
+const VOICED_TO_COMPOSED = new Map([
+  [0x304b, 0x304c], [0x304d, 0x304e], [0x304f, 0x3050], [0x3051, 0x3052],
+  [0x3053, 0x3054], [0x3055, 0x3056], [0x3057, 0x3058], [0x3059, 0x305a],
+  [0x305b, 0x305c], [0x305d, 0x305e], [0x305f, 0x3060], [0x3061, 0x3062],
+  [0x3064, 0x3065], [0x3066, 0x3067], [0x3068, 0x3069], [0x306f, 0x3070],
+  [0x3072, 0x3073], [0x3075, 0x3076], [0x3078, 0x3079], [0x307b, 0x307c],
+  [0x3046, 0x3094], [0x309d, 0x309e], [0x30ef, 0x30f7], [0x30f0, 0x30f8],
+  [0x30f1, 0x30f9], [0x30f2, 0x30fa], [0x30fd, 0x30fe]
+]);
+
+// prettier-ignore
+const SEMIVOICED_TO_COMPOSED = new Map([
+  [0x306f, 0x3071], [0x3072, 0x3074], [0x3075, 0x3077], [0x3078, 0x307a],
+  [0x307b, 0x307d],
+]);
+
+// Kana normalization
+//
+// Converts half-width katakana and full-width katakana to hiragana and also
+// converts decomposed characters to their composed equivalents.
 //
 // Returns the hiragana input and an array that maps each character
 // in the hiragana output to the corresponding length in the original input.
@@ -36,14 +45,15 @@ const SEMIVOICED_KATAKANA_TO_HIRAGANA = [
 // TODO: Translate this range https://en.wikipedia.org/wiki/Enclosed_Ideographic_Supplement
 //
 // TODO: Translate the first part of this range: https://en.wikipedia.org/wiki/CJK_Compatibility
-export function kanaToHiragana(input: string): [string, number[]] {
+export function normalizeKana(input: string): [string, number[]] {
   let inputLengths = [0];
-  let previous = 0;
+  let previousOriginal = 0;
   let result = '';
 
   for (let i = 0; i < input.length; ++i) {
     let originalChar = input.charCodeAt(i);
     let c = originalChar;
+    let lastChar = result.length ? result.charCodeAt(result.length - 1) : 0;
 
     // Full-width katakana to hiragana
     if ((c >= 0x30a1 && c <= 0x30f6) || c === 0x30fd || c === 0x30fe) {
@@ -51,35 +61,50 @@ export function kanaToHiragana(input: string): [string, number[]] {
     } else if (c >= 0xff65 && c <= 0xff9d) {
       // Half-width katakana to hiragana
       c = HANKAKU_KATAKANA_TO_HIRAGANA[c - 0xff65];
-    } else if (c == 0xff9e) {
-      // Voiced (used in half-width katakana) to hiragana
-      if (previous >= 0xff73 && previous <= 0xff8e) {
+    } else if (c === 0x3099 || c === 0xff9e) {
+      // Special case handling for: ヷヸヹヺ
+      //
+      // These characters cannot be converted into hiragana but if we encounter
+      // them in decomposed form, e.g. 0x30EF (ワ) + 0x3099 (゛), we will
+      // initially convert 0x30EF (ワ) into hiragana 0x308F (わ).
+      // When we encounter the 0x3099, however, we should recognize that the
+      // previous character was actually katakana, and produce 0x30F7 (ヷ)
+      // instead.
+      if (previousOriginal >= 0x30ef && previousOriginal <= 0x30f2) {
         result = result.slice(0, -1);
-        c = VOICED_KATAKANA_TO_HIRAGANA[previous - 0xff73];
+        c = previousOriginal + 8;
+      } else {
+        // Decomposed voiced mark (full-width or half-width)
+        const composed = VOICED_TO_COMPOSED.get(lastChar);
+        if (composed) {
+          result = result.slice(0, -1);
+          c = composed;
+        }
       }
-    } else if (c == 0xff9f) {
-      // Semi-voiced (used in half-width katakana) to hiragana
-      if (previous >= 0xff8a && previous <= 0xff8e) {
+    } else if (c === 0x309a || c === 0xff9f) {
+      // Decomposed semi-voiced mark (full-width or half-width)
+      const composed = SEMIVOICED_TO_COMPOSED.get(lastChar);
+      if (composed) {
         result = result.slice(0, -1);
-        c = SEMIVOICED_KATAKANA_TO_HIRAGANA[previous - 0xff8a];
+        c = composed;
       }
     } else if (c == 0xff5e && i > 0) {
       // Ignore ～ (but only if it's in the middle/end of a word)
-      previous = 0;
+      previousOriginal = 0;
       continue;
     }
 
     result += String.fromCharCode(c);
     // Need to keep real length because of the half-width semi/voiced conversion
     inputLengths[result.length] = i + 1;
-    previous = originalChar;
+    previousOriginal = originalChar;
   }
 
   return [result, inputLengths];
 }
 
 export function normalizeInput(input: string): [string, number[]] {
-  let [normalized, inputLengths] = kanaToHiragana(input);
+  let [normalized, inputLengths] = normalizeKana(input);
 
   // Truncate if we find characters outside the expected range.
   for (let i = 0; i < input.length; ++i) {